refactor: refactor presence to hook and use presence throughout the app
